// src/types/types.ts
export interface UserData {
  _id: string;
  username: string;
  fullname: string;
  email: string;
  phoneNumber: string;
  role: string;
}

export interface PersistedUserInfo {
  _id: string;
  role: string;
  username: string;
  fullname: string;
  email: string;
}